Refactor TerminalConsole
